Moved package of moving_avg
